pyproject.toml
[tool.poetry]
name = "etl-parquet-athena-glue"
version = "0.1.0"
description = ""
authors = ["omaucosta <mauricios.costta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.0"
pandas = "^2.2.2"
pyarrow = "^17.0.0"
python-dotenv = "^1.0.1"
update = "^0.0.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
